This is a squash commit for this issue on 29 May at 14:52 IST
