+++ test/e2e/config/multikueue/kubeflow-manifest-patch/kustomization.yaml
@@ -15,4 +15,3 @@ patches:
       group: apps
       kind: Deployment
       name: training-operator
-
