+++ test/e2e/multikueue/baseline/e2e_test.go
@@ -20,16 +20,10 @@ import (
 	"context"
 	"fmt"
 	"regexp"
-	"strconv"
 
 	"github.com/google/go-cmp/cmp/cmpopts"
-	kfmpi "github.com/kubeflow/mpi-operator/pkg/apis/kubeflow/v2beta1"
-	kftrainer "github.com/kubeflow/trainer/v2/pkg/apis/trainer/v1alpha1"
-	kftraining "github.com/kubeflow/training-operator/pkg/apis/kubeflow.org/v1"
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
-	awv1beta2 "github.com/project-codeflare/appwrapper/api/v1beta2"
-	rayv1 "github.com/ray-project/kuberay/ray-operator/apis/ray/v1"
 	appsv1 "k8s.io/api/apps/v1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -40,41 +34,21 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	jobset "sigs.k8s.io/jobset/api/jobset/v1alpha2"
-	leaderworkersetv1 "sigs.k8s.io/lws/api/leaderworkerset/v1"
 
 	kueue "sigs.k8s.io/kueue/apis/kueue/v1beta2"
 	"sigs.k8s.io/kueue/pkg/controller/constants"
-	workloadaw "sigs.k8s.io/kueue/pkg/controller/jobs/appwrapper"
 	workloadjob "sigs.k8s.io/kueue/pkg/controller/jobs/job"
-	workloadjobset "sigs.k8s.io/kueue/pkg/controller/jobs/jobset"
-	workloadpytorchjob "sigs.k8s.io/kueue/pkg/controller/jobs/kubeflow/jobs/pytorchjob"
-	workloadleaderworkerset "sigs.k8s.io/kueue/pkg/controller/jobs/leaderworkerset"
-	workloadmpijob "sigs.k8s.io/kueue/pkg/controller/jobs/mpijob"
 	workloadpod "sigs.k8s.io/kueue/pkg/controller/jobs/pod"
 	podconstants "sigs.k8s.io/kueue/pkg/controller/jobs/pod/constants"
-	workloadraycluster "sigs.k8s.io/kueue/pkg/controller/jobs/raycluster"
-	workloadrayjob "sigs.k8s.io/kueue/pkg/controller/jobs/rayjob"
-	workloadrayservice "sigs.k8s.io/kueue/pkg/controller/jobs/rayservice"
 	workloadstatefulset "sigs.k8s.io/kueue/pkg/controller/jobs/statefulset"
-	workloadtrainjob "sigs.k8s.io/kueue/pkg/controller/jobs/trainjob"
 	"sigs.k8s.io/kueue/pkg/util/admissioncheck"
 	utilpod "sigs.k8s.io/kueue/pkg/util/pod"
 	utiltesting "sigs.k8s.io/kueue/pkg/util/testing"
 	utiltestingapi "sigs.k8s.io/kueue/pkg/util/testing/v1beta2"
-	testingaw "sigs.k8s.io/kueue/pkg/util/testingjobs/appwrapper"
 	testingdeployment "sigs.k8s.io/kueue/pkg/util/testingjobs/deployment"
 	testingjob "sigs.k8s.io/kueue/pkg/util/testingjobs/job"
-	testingjobset "sigs.k8s.io/kueue/pkg/util/testingjobs/jobset"
-	testingleaderworkerset "sigs.k8s.io/kueue/pkg/util/testingjobs/leaderworkerset"
-	testingmpijob "sigs.k8s.io/kueue/pkg/util/testingjobs/mpijob"
 	testingpod "sigs.k8s.io/kueue/pkg/util/testingjobs/pod"
-	testingpytorchjob "sigs.k8s.io/kueue/pkg/util/testingjobs/pytorchjob"
-	testingraycluster "sigs.k8s.io/kueue/pkg/util/testingjobs/raycluster"
-	testingrayjob "sigs.k8s.io/kueue/pkg/util/testingjobs/rayjob"
-	testingrayservice "sigs.k8s.io/kueue/pkg/util/testingjobs/rayservice"
 	testingstatefulset "sigs.k8s.io/kueue/pkg/util/testingjobs/statefulset"
-	testingtrainjob "sigs.k8s.io/kueue/pkg/util/testingjobs/trainjob"
 	"sigs.k8s.io/kueue/pkg/workload"
 	"sigs.k8s.io/kueue/test/util"
 )
@@ -246,17 +220,6 @@ var _ = ginkgo.Describe("MultiKueue", func() {
 	})
 
 	ginkgo.AfterEach(func() {
-		// Clean up resources created by the RayService test on all clusters.
-		rayServiceConfigMap := &corev1.ConfigMap{ObjectMeta: metav1.ObjectMeta{Name: "rayservice-hello", Namespace: managerNs.Name}}
-		gomega.Expect(client.IgnoreNotFound(k8sManagerClient.Delete(ctx, rayServiceConfigMap))).To(gomega.Succeed())
-		gomega.Expect(client.IgnoreNotFound(k8sWorker1Client.Delete(ctx, rayServiceConfigMap.DeepCopy()))).To(gomega.Succeed())
-		gomega.Expect(client.IgnoreNotFound(k8sWorker2Client.Delete(ctx, rayServiceConfigMap.DeepCopy()))).To(gomega.Succeed())
-
-		rayService := &rayv1.RayService{ObjectMeta: metav1.ObjectMeta{Name: "rayservice1", Namespace: managerNs.Name}}
-		gomega.Expect(client.IgnoreNotFound(k8sManagerClient.Delete(ctx, rayService))).To(gomega.Succeed())
-		gomega.Expect(client.IgnoreNotFound(k8sWorker1Client.Delete(ctx, rayService.DeepCopy()))).To(gomega.Succeed())
-		gomega.Expect(client.IgnoreNotFound(k8sWorker2Client.Delete(ctx, rayService.DeepCopy()))).To(gomega.Succeed())
-
 		gomega.Expect(util.DeleteNamespace(ctx, k8sManagerClient, managerNs)).To(gomega.Succeed())
 		gomega.Expect(util.DeleteNamespace(ctx, k8sWorker1Client, worker1Ns)).To(gomega.Succeed())
 		gomega.Expect(util.DeleteNamespace(ctx, k8sWorker2Client, worker2Ns)).To(gomega.Succeed())
@@ -522,173 +485,6 @@ var _ = ginkgo.Describe("MultiKueue", func() {
 			})
 		})
 
-		ginkgo.It("Should sync a LeaderWorkerSet and run replicas on worker cluster", func() {
-			lws := testingleaderworkerset.MakeLeaderWorkerSet("leaderworkerset", managerNs.Name).
-				Image(util.GetAgnHostImage(), util.BehaviorWaitForDeletion).
-				Replicas(2).
-				Size(2).
-				RequestAndLimit(corev1.ResourceCPU, "100m").
-				RequestAndLimit(corev1.ResourceMemory, "100M").
-				Queue(managerLq.Name).
-				TerminationGracePeriod(1).
-				Obj()
-
-			ginkgo.By("Creating the leaderworkerset", func() {
-				util.MustCreate(ctx, k8sManagerClient, lws)
-			})
-
-			createdLWS := &leaderworkersetv1.LeaderWorkerSet{}
-			gomega.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(lws), createdLWS)).To(gomega.Succeed())
-
-			wlLookupKey0 := types.NamespacedName{
-				Name:      workloadleaderworkerset.GetWorkloadName(createdLWS.UID, createdLWS.Name, "0"),
-				Namespace: managerNs.Name,
-			}
-			wlLookupKey1 := types.NamespacedName{
-				Name:      workloadleaderworkerset.GetWorkloadName(createdLWS.UID, createdLWS.Name, "1"),
-				Namespace: managerNs.Name,
-			}
-
-			admittedWorkerName := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlLookupKey0, multiKueueAc.Name)
-			workerClient := kubernetesClients[admittedWorkerName].client
-
-			ginkgo.By("Verifying both workloads are admitted on the same worker", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					wl0 := &kueue.Workload{}
-					g.Expect(workerClient.Get(ctx, wlLookupKey0, wl0)).To(gomega.Succeed())
-					g.Expect(workload.IsAdmitted(wl0)).To(gomega.BeTrue())
-					wl1 := &kueue.Workload{}
-					g.Expect(workerClient.Get(ctx, wlLookupKey1, wl1)).To(gomega.Succeed())
-					g.Expect(workload.IsAdmitted(wl1)).To(gomega.BeTrue())
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Waiting for LWS to be synced to worker cluster", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					workerLWS := &leaderworkersetv1.LeaderWorkerSet{}
-					g.Expect(workerClient.Get(ctx, client.ObjectKeyFromObject(lws), workerLWS)).To(gomega.Succeed())
-				}, util.Timeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Waiting for all replicas to be ready on worker cluster", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					workerLWS := &leaderworkersetv1.LeaderWorkerSet{}
-					g.Expect(workerClient.Get(ctx, client.ObjectKeyFromObject(lws), workerLWS)).To(gomega.Succeed())
-					g.Expect(workerLWS.Status.ReadyReplicas).To(gomega.Equal(int32(2)))
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Verifying pods on management cluster remain gated", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					pods := &corev1.PodList{}
-					g.Expect(k8sManagerClient.List(ctx, pods, client.InNamespace(managerNs.Name), client.MatchingLabels{
-						leaderworkersetv1.SetNameLabelKey: lws.Name,
-					})).To(gomega.Succeed())
-					g.Expect(pods.Items).ToNot(gomega.BeEmpty())
-					for _, pod := range pods.Items {
-						g.Expect(utilpod.HasGate(&pod, podconstants.SchedulingGateName)).To(gomega.BeTrue())
-					}
-				}, util.Timeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Deleting the leaderworkerset", func() {
-				util.ExpectObjectToBeDeleted(ctx, k8sManagerClient, lws, true)
-				util.ExpectObjectToBeDeletedWithTimeout(ctx, workerClient, lws, false, util.MediumTimeout)
-			})
-
-			ginkgo.By("Checking that all workloads are deleted from manager and worker clusters", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					g.Expect(k8sManagerClient.Get(ctx, wlLookupKey0, &kueue.Workload{})).To(utiltesting.BeNotFoundError())
-					g.Expect(k8sManagerClient.Get(ctx, wlLookupKey1, &kueue.Workload{})).To(utiltesting.BeNotFoundError())
-					g.Expect(workerClient.Get(ctx, wlLookupKey0, &kueue.Workload{})).To(utiltesting.BeNotFoundError())
-					g.Expect(workerClient.Get(ctx, wlLookupKey1, &kueue.Workload{})).To(utiltesting.BeNotFoundError())
-				}, util.Timeout, util.Interval).Should(gomega.Succeed())
-			})
-		})
-
-		ginkgo.It("Should dispatch all LeaderWorkerSet workloads to the same worker", func() {
-			const lwsReplicas = 3
-			lws := testingleaderworkerset.MakeLeaderWorkerSet("leaderworkerset", managerNs.Name).
-				Image(util.GetAgnHostImage(), util.BehaviorWaitForDeletion).
-				Replicas(lwsReplicas).
-				Size(2).
-				RequestAndLimit(corev1.ResourceCPU, "100m").
-				RequestAndLimit(corev1.ResourceMemory, "100M").
-				Queue(managerLq.Name).
-				TerminationGracePeriod(1).
-				Obj()
-
-			ginkgo.By("Creating the leaderworkerset", func() {
-				util.MustCreate(ctx, k8sManagerClient, lws)
-			})
-
-			createdLWS := &leaderworkersetv1.LeaderWorkerSet{}
-			gomega.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(lws), createdLWS)).To(gomega.Succeed())
-
-			wlKeys := make([]types.NamespacedName, lwsReplicas)
-			for i := range lwsReplicas {
-				wlKeys[i] = types.NamespacedName{
-					Name:      workloadleaderworkerset.GetWorkloadName(createdLWS.UID, createdLWS.Name, strconv.Itoa(i)),
-					Namespace: managerNs.Name,
-				}
-			}
-
-			ginkgo.By("Waiting for workloads to be created on manager cluster", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					for _, key := range wlKeys {
-						wl := &kueue.Workload{}
-						g.Expect(k8sManagerClient.Get(ctx, key, wl)).To(gomega.Succeed())
-					}
-				}, util.Timeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			admittedWorkerName := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlKeys[0], multiKueueAc.Name)
-			workerClient := kubernetesClients[admittedWorkerName].client
-
-			ginkgo.By("Verifying primary workload is admitted on worker2", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					wl := &kueue.Workload{}
-					g.Expect(workerClient.Get(ctx, wlKeys[0], wl)).To(gomega.Succeed())
-					g.Expect(workload.IsAdmitted(wl)).To(gomega.BeTrue())
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Verifying LWS is synced to worker cluster", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					workerLWS := &leaderworkersetv1.LeaderWorkerSet{}
-					g.Expect(workerClient.Get(ctx, client.ObjectKeyFromObject(lws), workerLWS)).To(gomega.Succeed())
-				}, util.Timeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Verifying follower workloads are dispatched to the same worker cluster", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					primaryWl := &kueue.Workload{}
-					g.Expect(k8sManagerClient.Get(ctx, wlKeys[0], primaryWl)).To(gomega.Succeed())
-					g.Expect(primaryWl.Status.ClusterName).ToNot(gomega.BeNil())
-					for _, key := range wlKeys[1:] {
-						wl := &kueue.Workload{}
-						g.Expect(k8sManagerClient.Get(ctx, key, wl)).To(gomega.Succeed())
-						g.Expect(wl.Status.ClusterName).ToNot(gomega.BeNil())
-						g.Expect(*wl.Status.ClusterName).To(gomega.Equal(*primaryWl.Status.ClusterName))
-					}
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Deleting the leaderworkerset", func() {
-				util.ExpectObjectToBeDeleted(ctx, k8sManagerClient, lws, true)
-				util.ExpectObjectToBeDeletedWithTimeout(ctx, workerClient, lws, false, util.MediumTimeout)
-			})
-
-			ginkgo.By("Checking that all workloads are deleted from manager and worker clusters", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					for _, key := range wlKeys {
-						g.Expect(k8sManagerClient.Get(ctx, key, &kueue.Workload{})).To(utiltesting.BeNotFoundError())
-						g.Expect(workerClient.Get(ctx, key, &kueue.Workload{})).To(utiltesting.BeNotFoundError())
-					}
-				}, util.Timeout, util.Interval).Should(gomega.Succeed())
-			})
-		})
-
 		ginkgo.It("Should create a pod group on worker if admitted", func() {
 			numPods := 2
 			groupName := "test-group"
@@ -824,477 +620,8 @@ var _ = ginkgo.Describe("MultiKueue", func() {
 					cmpopts.IgnoreFields(batchv1.JobCondition{}, "LastTransitionTime", "LastProbeTime", "Reason", "Message"))))
 			})
 		})
-
-		ginkgo.It("Should run a jobSet on worker if admitted", func() {
-			jobSet := testingjobset.MakeJobSet("job-set", managerNs.Name).
-				Queue(managerLq.Name).
-				ReplicatedJobs(
-					testingjobset.ReplicatedJobRequirements{
-						Name:        "replicated-job-1",
-						Replicas:    2,
-						Parallelism: 2,
-						Completions: 2,
-						Image:       util.GetAgnHostImage(),
-						// Give it the time to be observed Active in the live status update step.
-						Args: util.BehaviorWaitForDeletion,
-					},
-				).
-				RequestAndLimit("replicated-job-1", corev1.ResourceCPU, "100m").
-				RequestAndLimit("replicated-job-1", corev1.ResourceMemory, "100M").
-				TerminationGracePeriod(1).
-				Obj()
-
-			ginkgo.By("Creating the jobSet", func() {
-				util.MustCreate(ctx, k8sManagerClient, jobSet)
-			})
-
-			createdLeaderWorkload := &kueue.Workload{}
-			wlLookupKey := types.NamespacedName{Name: workloadjobset.GetWorkloadNameForJobSet(jobSet.Name, jobSet.UID), Namespace: managerNs.Name}
-
-			admittedWorkerName := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlLookupKey, multiKueueAc.Name)
-			admittedWorker := kubernetesClients[admittedWorkerName]
-
-			ginkgo.By("Waiting for the jobSet to get status updates", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					createdJobset := &jobset.JobSet{}
-					g.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(jobSet), createdJobset)).To(gomega.Succeed())
-
-					g.Expect(createdJobset.Status.ReplicatedJobsStatus).To(gomega.BeComparableTo([]jobset.ReplicatedJobStatus{
-						{
-							Name:   "replicated-job-1",
-							Ready:  2,
-							Active: 2,
-						},
-					}, cmpopts.IgnoreFields(jobset.ReplicatedJobStatus{}, "Succeeded", "Failed")))
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Finishing the jobset pods", func() {
-				listOpts := util.GetListOptsFromLabel(fmt.Sprintf("jobset.sigs.k8s.io/jobset-name=%s", jobSet.Name))
-				util.WaitForActivePodsAndTerminate(ctx, admittedWorker.client, admittedWorker.restClient, admittedWorker.cfg, jobSet.Namespace, 4, 0, listOpts)
-			})
-
-			ginkgo.By("Waiting for the jobSet to finish", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					g.Expect(k8sManagerClient.Get(ctx, wlLookupKey, createdLeaderWorkload)).To(gomega.Succeed())
-
-					g.Expect(apimeta.FindStatusCondition(createdLeaderWorkload.Status.Conditions, kueue.WorkloadFinished)).To(gomega.BeComparableTo(&metav1.Condition{
-						Type:    kueue.WorkloadFinished,
-						Status:  metav1.ConditionTrue,
-						Reason:  kueue.WorkloadFinishedReasonSucceeded,
-						Message: "jobset completed successfully",
-					}, util.IgnoreConditionTimestampsAndObservedGeneration))
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Checking no objects are left in the worker clusters and the jobSet is completed", func() {
-				util.ExpectObjectToBeDeletedOnClusters(ctx, createdLeaderWorkload, k8sWorker1Client, k8sWorker2Client)
-				util.ExpectObjectToBeDeletedOnClusters(ctx, jobSet, k8sWorker1Client, k8sWorker2Client)
-
-				createdJobSet := &jobset.JobSet{}
-				gomega.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(jobSet), createdJobSet)).To(gomega.Succeed())
-				gomega.Expect(ptr.Deref(createdJobSet.Spec.Suspend, true)).To(gomega.BeFalse())
-				gomega.Expect(createdJobSet.Status.Conditions).To(gomega.ContainElement(gomega.BeComparableTo(
-					metav1.Condition{
-						Type:    string(jobset.JobSetCompleted),
-						Status:  metav1.ConditionTrue,
-						Reason:  "AllJobsCompleted",
-						Message: "jobset completed successfully",
-					},
-					util.IgnoreConditionTimestampsAndObservedGeneration)))
-			})
-		})
-
-		ginkgo.It("Should run an appwrapper containing a job on worker if admitted", func() {
-			jobName := "job-1"
-			aw := testingaw.MakeAppWrapper("aw", managerNs.Name).
-				Queue(managerLq.Name).
-				Component(testingaw.Component{
-					Template: testingjob.MakeJob(jobName, managerNs.Name).
-						SetTypeMeta().
-						Suspend(false).
-						Image(util.GetAgnHostImage(), util.BehaviorWaitForDeletion). // Give it the time to be observed Active in the live status update step.
-						Parallelism(2).
-						RequestAndLimit(corev1.ResourceCPU, "100m").
-						RequestAndLimit(corev1.ResourceMemory, "100M").
-						TerminationGracePeriod(1).
-						SetTypeMeta().Obj(),
-				}).
-				Obj()
-
-			ginkgo.By("Creating the appwrapper", func() {
-				util.MustCreate(ctx, k8sManagerClient, aw)
-			})
-
-			wlLookupKey := types.NamespacedName{Name: workloadaw.GetWorkloadNameForAppWrapper(aw.Name, aw.UID), Namespace: managerNs.Name}
-
-			admittedWorkerName := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlLookupKey, multiKueueAc.Name)
-			admittedWorker := kubernetesClients[admittedWorkerName]
-
-			ginkgo.By("Waiting for the appwrapper to get status updates", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					createdAppWrapper := &awv1beta2.AppWrapper{}
-					g.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(aw), createdAppWrapper)).To(gomega.Succeed())
-					g.Expect(createdAppWrapper.Status.Phase).To(gomega.Equal(awv1beta2.AppWrapperRunning))
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("Finishing the wrapped job's pods", func() {
-				listOpts := util.GetListOptsFromLabel(fmt.Sprintf("batch.kubernetes.io/job-name=%s", jobName))
-				util.WaitForActivePodsAndTerminate(ctx, admittedWorker.client, admittedWorker.restClient, admittedWorker.cfg, aw.Namespace, 2, 0, listOpts)
-			})
-
-			ginkgo.By("Waiting for the appwrapper to finish", func() {
-				util.ExpectWorkloadToFinish(ctx, k8sManagerClient, wlLookupKey)
-			})
-
-			ginkgo.By("Checking no objects are left in the worker clusters and the appwrapper is completed", func() {
-				createdWorkload := &kueue.Workload{}
-				gomega.Expect(k8sManagerClient.Get(ctx, wlLookupKey, createdWorkload)).To(gomega.Succeed())
-				util.ExpectObjectToBeDeletedOnClusters(ctx, createdWorkload, k8sWorker1Client, k8sWorker2Client)
-				util.ExpectObjectToBeDeletedOnClusters(ctx, aw, k8sWorker1Client, k8sWorker2Client)
-
-				createdAppWrapper := &awv1beta2.AppWrapper{}
-				gomega.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(aw), createdAppWrapper)).To(gomega.Succeed())
-				gomega.Expect(createdAppWrapper.Spec.Suspend).To(gomega.BeFalse())
-				gomega.Expect(createdAppWrapper.Status.Phase).To(gomega.Equal(awv1beta2.AppWrapperSucceeded))
-			})
-		})
-
-		ginkgo.It("Should run a kubeflow PyTorchJob on worker if admitted", func() {
-			pyTorchJob := testingpytorchjob.MakePyTorchJob("pytorchjob1", managerNs.Name).
-				ManagedBy(kueue.MultiKueueControllerName).
-				Queue(managerLq.Name).
-				PyTorchReplicaSpecs(
-					testingpytorchjob.PyTorchReplicaSpecRequirement{
-						ReplicaType:   kftraining.PyTorchJobReplicaTypeMaster,
-						ReplicaCount:  1,
-						RestartPolicy: "Never",
-						Image:         util.GetAgnHostImage(),
-						Args:          util.BehaviorExitFast,
-					},
-				).
-				RequestAndLimit(kftraining.PyTorchJobReplicaTypeMaster, corev1.ResourceCPU, "100m").
-				RequestAndLimit(kftraining.PyTorchJobReplicaTypeMaster, corev1.ResourceMemory, "100M").
-				RequestAndLimit(kftraining.PyTorchJobReplicaTypeWorker, corev1.ResourceCPU, "100m").
-				RequestAndLimit(kftraining.PyTorchJobReplicaTypeWorker, corev1.ResourceMemory, "100M").
-				Obj()
-
-			ginkgo.By("Creating the PyTorchJob", func() {
-				util.MustCreate(ctx, k8sManagerClient, pyTorchJob)
-			})
-
-			wlLookupKey := types.NamespacedName{Name: workloadpytorchjob.GetWorkloadNameForPyTorchJob(pyTorchJob.Name, pyTorchJob.UID), Namespace: managerNs.Name}
-
-			admittedWorker := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlLookupKey, multiKueueAc.Name)
-			ginkgo.GinkgoLogr.Info("PyTorchJob %s is admitted in worker cluster %s", pyTorchJob.Name, admittedWorker)
-
-			ginkgo.By("Waiting for the PyTorchJob to finish", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					createdPyTorchJob := &kftraining.PyTorchJob{}
-					g.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(pyTorchJob), createdPyTorchJob)).To(gomega.Succeed())
-					g.Expect(createdPyTorchJob.Status.ReplicaStatuses[kftraining.PyTorchJobReplicaTypeMaster]).To(gomega.BeComparableTo(
-						&kftraining.ReplicaStatus{
-							Active:    0,
-							Succeeded: 1,
-							Selector: fmt.Sprintf(
-								"training.kubeflow.org/job-name=%s,training.kubeflow.org/operator-name=pytorchjob-controller,training.kubeflow.org/replica-type=master",
-								createdPyTorchJob.Name,
-							),
-						},
-					))
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-				util.ExpectWorkloadToFinish(ctx, k8sManagerClient, wlLookupKey)
-			})
-
-			ginkgo.By("Checking no objects are left in the worker clusters and the PyTorchJob is completed", func() {
-				wl := &kueue.Workload{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      wlLookupKey.Name,
-						Namespace: wlLookupKey.Namespace,
-					},
-				}
-				util.ExpectObjectToBeDeletedOnClusters(ctx, wl, k8sWorker1Client, k8sWorker2Client)
-				util.ExpectObjectToBeDeletedOnClusters(ctx, pyTorchJob, k8sWorker1Client, k8sWorker2Client)
-			})
-		})
-
-		ginkgo.It("Should run a MPIJob on worker if admitted", func() {
-			mpijob := testingmpijob.MakeMPIJob("mpijob1", managerNs.Name).
-				Queue(managerLq.Name).
-				ManagedBy(kueue.MultiKueueControllerName).
-				MPIJobReplicaSpecs(
-					testingmpijob.MPIJobReplicaSpecRequirement{
-						ReplicaType:   kfmpi.MPIReplicaTypeLauncher,
-						ReplicaCount:  1,
-						RestartPolicy: "OnFailure",
-						Image:         util.GetAgnHostImage(),
-						Args:          util.BehaviorExitFast,
-					},
-					testingmpijob.MPIJobReplicaSpecRequirement{
-						ReplicaType:   kfmpi.MPIReplicaTypeWorker,
-						ReplicaCount:  1,
-						RestartPolicy: "OnFailure",
-						Image:         util.GetAgnHostImage(),
-						Args:          util.BehaviorExitFast,
-					},
-				).
-				RequestAndLimit(kfmpi.MPIReplicaTypeLauncher, corev1.ResourceCPU, "100m").
-				RequestAndLimit(kfmpi.MPIReplicaTypeLauncher, corev1.ResourceMemory, "100M").
-				RequestAndLimit(kfmpi.MPIReplicaTypeWorker, corev1.ResourceCPU, "100m").
-				RequestAndLimit(kfmpi.MPIReplicaTypeWorker, corev1.ResourceMemory, "100M").
-				Obj()
-
-			ginkgo.By("Creating the MPIJob", func() {
-				util.MustCreate(ctx, k8sManagerClient, mpijob)
-			})
-
-			wlLookupKey := types.NamespacedName{Name: workloadmpijob.GetWorkloadNameForMPIJob(mpijob.Name, mpijob.UID), Namespace: managerNs.Name}
-
-			admittedWorker := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlLookupKey, multiKueueAc.Name)
-			ginkgo.GinkgoLogr.Info("MPIJob %s is admitted in worker cluster %s", mpijob.Name, admittedWorker)
-
-			ginkgo.By("Waiting for the MPIJob to finish", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					createdMPIJob := &kfmpi.MPIJob{}
-					g.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(mpijob), createdMPIJob)).To(gomega.Succeed())
-					g.Expect(createdMPIJob.Status.ReplicaStatuses[kfmpi.MPIReplicaTypeLauncher]).To(gomega.BeComparableTo(
-						&kfmpi.ReplicaStatus{
-							Active:    0,
-							Succeeded: 1,
-						},
-					))
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-				util.ExpectWorkloadToFinish(ctx, k8sManagerClient, wlLookupKey)
-			})
-
-			ginkgo.By("Checking no objects are left in the worker clusters and the MPIJob is completed", func() {
-				wl := &kueue.Workload{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      wlLookupKey.Name,
-						Namespace: wlLookupKey.Namespace,
-					},
-				}
-				util.ExpectObjectToBeDeletedOnClusters(ctx, wl, k8sWorker1Client, k8sWorker2Client)
-				util.ExpectObjectToBeDeletedOnClusters(ctx, mpijob, k8sWorker1Client, k8sWorker2Client)
-			})
-		})
-
-		ginkgo.It("Should run a TrainJob on worker if admitted", func() {
-			trainjob := testingtrainjob.MakeTrainJob("trainjob-test", managerNs.Name).
-				RuntimeRefName("torch-distributed").
-				Queue(managerLq.Name).
-				RequestAndLimit(corev1.ResourceCPU, "100m", "100m").
-				RequestAndLimit(corev1.ResourceMemory, "100M", "100M").
-				// Even if we override the image coming from the TrainingRuntime, we still need to set the command and args
-				TrainerImage(util.GetAgnHostImage(), []string{"/agnhost"}, util.BehaviorExitFast).
-				Obj()
-
-			ginkgo.By("Creating the trainjob", func() {
-				util.MustCreate(ctx, k8sManagerClient, trainjob)
-			})
-
-			wlLookupKey := types.NamespacedName{Name: workloadtrainjob.GetWorkloadNameForTrainJob(trainjob.Name, trainjob.UID), Namespace: managerNs.Name}
-
-			admittedWorker := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlLookupKey, multiKueueAc.Name)
-			ginkgo.GinkgoLogr.Info("TrainJob %s is admitted in worker cluster %s", trainjob.Name, admittedWorker)
-
-			ginkgo.By("Checking the TrainJob is ready", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					createdTrainJob := &kftrainer.TrainJob{}
-					g.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(trainjob), createdTrainJob)).To(gomega.Succeed())
-					g.Expect(ptr.Deref(createdTrainJob.Spec.Suspend, false)).To(gomega.BeFalse())
-				}, util.VeryLongTimeout, util.Interval).Should(gomega.Succeed())
-			})
-		})
-
-		ginkgo.When("Ray integration tests", ginkgo.Ordered, func() {
-			ginkgo.It("Should run a RayJob on worker if admitted", func() {
-				kuberayTestImage := util.GetKuberayTestImage()
-				rayjob := testingrayjob.MakeJob("rayjob1", managerNs.Name).
-					Suspend(true).
-					Queue(managerLq.Name).
-					WithSubmissionMode(rayv1.K8sJobMode).
-					Entrypoint("python -c \"import ray; ray.init(); print(ray.cluster_resources())\"").
-					RequestAndLimit(rayv1.HeadNode, corev1.ResourceCPU, "1").
-					RequestAndLimit(rayv1.WorkerNode, corev1.ResourceCPU, "0.5").
-					Image(rayv1.HeadNode, kuberayTestImage).
-					Image(rayv1.WorkerNode, kuberayTestImage).
-					Obj()
-
-				ginkgo.By("Creating the RayJob", func() {
-					util.MustCreate(ctx, k8sManagerClient, rayjob)
-				})
-
-				wlLookupKey := types.NamespacedName{Name: workloadrayjob.GetWorkloadNameForRayJob(rayjob.Name, rayjob.UID), Namespace: managerNs.Name}
-
-				admittedWorker := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlLookupKey, multiKueueAc.Name)
-				ginkgo.GinkgoLogr.Info(fmt.Sprintf("RayJob %s/%s is admitted in worker cluster %s", rayjob.Name, rayjob.Namespace, admittedWorker))
-
-				ginkgo.By("Waiting for the RayJob to finish", func() {
-					gomega.Eventually(func(g gomega.Gomega) {
-						createdRayJob := &rayv1.RayJob{}
-						g.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(rayjob), createdRayJob)).To(gomega.Succeed())
-						g.Expect(createdRayJob.Status.JobDeploymentStatus).To(gomega.Equal(rayv1.JobDeploymentStatusComplete))
-					}, util.VeryLongTimeout, util.Interval).Should(gomega.Succeed())
-					util.ExpectWorkloadToFinish(ctx, k8sManagerClient, wlLookupKey)
-				})
-
-				ginkgo.By("Checking no objects are left in the worker clusters and the RayJob is completed", func() {
-					wl := &kueue.Workload{
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      wlLookupKey.Name,
-							Namespace: wlLookupKey.Namespace,
-						},
-					}
-					util.ExpectObjectToBeDeletedOnClusters(ctx, wl, k8sWorker1Client, k8sWorker2Client)
-					util.ExpectObjectToBeDeletedOnClusters(ctx, rayjob, k8sWorker1Client, k8sWorker2Client)
-				})
-			})
-
-			ginkgo.It("Should run a RayCluster on worker if admitted", func() {
-				kuberayTestImage := util.GetKuberayTestImage()
-				raycluster := testingraycluster.MakeCluster("raycluster1", managerNs.Name).
-					Suspend(true).
-					Queue(managerLq.Name).
-					RequestAndLimit(rayv1.HeadNode, corev1.ResourceCPU, "1").
-					RequestAndLimit(rayv1.WorkerNode, corev1.ResourceCPU, "0.5").
-					Image(rayv1.HeadNode, kuberayTestImage, []string{}).
-					Image(rayv1.WorkerNode, kuberayTestImage, []string{}).
-					Obj()
-
-				ginkgo.By("Creating the RayCluster", func() {
-					util.MustCreate(ctx, k8sManagerClient, raycluster)
-				})
-
-				wlLookupKey := types.NamespacedName{Name: workloadraycluster.GetWorkloadNameForRayCluster(raycluster.Name, raycluster.UID), Namespace: managerNs.Name}
-				// the execution should be given to the worker1
-				admittedWorker := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlLookupKey, multiKueueAc.Name)
-				ginkgo.GinkgoLogr.Info(fmt.Sprintf("RayCluster %s/%s is admitted in worker cluster %s", raycluster.Name, raycluster.Namespace, admittedWorker))
-
-				ginkgo.By("Checking the RayCluster is ready", func() {
-					gomega.Eventually(func(g gomega.Gomega) {
-						createdRayCluster := &rayv1.RayCluster{}
-						g.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(raycluster), createdRayCluster)).To(gomega.Succeed())
-						g.Expect(createdRayCluster.Status.DesiredWorkerReplicas).To(gomega.Equal(int32(1)))
-						g.Expect(createdRayCluster.Status.ReadyWorkerReplicas).To(gomega.Equal(int32(1)))
-						g.Expect(createdRayCluster.Status.AvailableWorkerReplicas).To(gomega.Equal(int32(1)))
-					}, util.VeryLongTimeout, util.Interval).Should(gomega.Succeed())
-				})
-			})
-
-			ginkgo.It("Should run a RayService on worker if admitted", func() {
-				kuberayTestImage := util.GetKuberayTestImage()
-
-				// Create ConfigMap with a simple Ray Serve application
-				configMap := &corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "rayservice-hello",
-						Namespace: managerNs.Name,
-					},
-					Data: map[string]string{
-						"hello_serve.py": `from ray import serve
-
-@serve.deployment
-class HelloWorld:
-    def __call__(self, request):
-        return "Hello, World!"
-
-app = HelloWorld.bind()`,
-					},
-				}
-
-				serveConfigV2 := `applications:
-  - name: hello_app
-    import_path: hello_serve:app
-    route_prefix: /
-    deployments:
-      - name: HelloWorld
-        num_replicas: 1
-        max_replicas_per_node: 1
-        ray_actor_options:
-          num_cpus: 0.2`
-
-				volumes := []corev1.Volume{
-					{
-						Name: "code-sample",
-						VolumeSource: corev1.VolumeSource{
-							ConfigMap: &corev1.ConfigMapVolumeSource{
-								LocalObjectReference: corev1.LocalObjectReference{
-									Name: "rayservice-hello",
-								},
-								Items: []corev1.KeyToPath{
-									{
-										Key:  "hello_serve.py",
-										Path: "hello_serve.py",
-									},
-								},
-							},
-						},
-					},
-				}
-				volumeMounts := []corev1.VolumeMount{
-					{
-						Name:      "code-sample",
-						MountPath: "/home/ray/samples",
-					},
-				}
-				env := []corev1.EnvVar{
-					{
-						Name:  "PYTHONPATH",
-						Value: "/home/ray/samples:$PYTHONPATH",
-					},
-				}
-
-				rayService := testingrayservice.MakeService("rayservice1", managerNs.Name).
-					Suspend(true).
-					Queue(managerLq.Name).
-					RequestAndLimit(rayv1.HeadNode, corev1.ResourceCPU, "1").
-					RequestAndLimit(rayv1.WorkerNode, corev1.ResourceCPU, "0.5").
-					Image(rayv1.HeadNode, kuberayTestImage).
-					Image(rayv1.WorkerNode, kuberayTestImage).
-					RayStartParam(rayv1.HeadNode, "object-store-memory", "100000000").
-					WithServeConfigV2(serveConfigV2).
-					Env(rayv1.HeadNode, env).
-					Env(rayv1.WorkerNode, env).
-					Volumes(rayv1.HeadNode, volumes).
-					Volumes(rayv1.WorkerNode, volumes).
-					VolumeMounts(rayv1.HeadNode, volumeMounts).
-					VolumeMounts(rayv1.WorkerNode, volumeMounts).
-					Obj()
-
-				rayService.Spec.RayClusterSpec.WorkerGroupSpecs[0].GroupName = "small-group"
-				rayService.Spec.RayClusterSpec.WorkerGroupSpecs[0].MinReplicas = ptr.To[int32](1)
-				rayService.Spec.RayClusterSpec.WorkerGroupSpecs[0].MaxReplicas = ptr.To[int32](2)
-
-				ginkgo.By("Creating the ConfigMap on all clusters", func() {
-					worker1ConfigMap := configMap.DeepCopy()
-					worker2ConfigMap := configMap.DeepCopy()
-					util.MustCreate(ctx, k8sManagerClient, configMap)
-					util.MustCreate(ctx, k8sWorker1Client, worker1ConfigMap)
-					util.MustCreate(ctx, k8sWorker2Client, worker2ConfigMap)
-				})
-
-				ginkgo.By("Creating the RayService", func() {
-					util.MustCreate(ctx, k8sManagerClient, rayService)
-				})
-
-				wlLookupKey := types.NamespacedName{Name: workloadrayservice.GetWorkloadNameForRayService(rayService.Name, rayService.UID), Namespace: managerNs.Name}
-
-				admittedWorker := ExpectWorkloadsToBeAdmittedAndGetWorkerName(ctx, k8sManagerClient, wlLookupKey, multiKueueAc.Name)
-				ginkgo.GinkgoLogr.Info(fmt.Sprintf("RayService %s/%s is admitted in worker cluster %s", rayService.Name, rayService.Namespace, admittedWorker))
-
-				ginkgo.By("Checking the RayService is running", func() {
-					gomega.Eventually(func(g gomega.Gomega) {
-						createdRayService := &rayv1.RayService{}
-						g.Expect(k8sManagerClient.Get(ctx, client.ObjectKeyFromObject(rayService), createdRayService)).To(gomega.Succeed())
-						g.Expect(createdRayService.Spec.RayClusterSpec.Suspend).To(gomega.Equal(new(false)))
-						g.Expect(apimeta.IsStatusConditionTrue(createdRayService.Status.Conditions, string(rayv1.RayServiceReady))).To(gomega.BeTrue())
-					}, util.VeryLongTimeout, util.Interval).Should(gomega.Succeed())
-				})
-			})
-		})
 	})
+
 	ginkgo.When("Preemption with a multikueue admission check", func() {
 		ginkgo.It("Should preempt a running low-priority workload when a high-priority workload is admitted (same worker)", func() {
 			lowJob := testingjob.MakeJob("", managerNs.Name).
